fixup! [cr145] Use `CopyFromSurfaceResult` callback

+++ components/ai_chat/content/browser/page_content_fetcher.cc
@@ -470,7 +470,7 @@ void OnScreenshot(FetchPageContentCallback callback,
                   const content::CopyFromSurfaceResult& result) {
   if (!result.has_value()) {
     LOG(ERROR) << "RenderWidgetHostView::CopyFromSurface failed with: "
-               << result.error();
+               << static_cast<int>(result.error());
     std::move(callback).Run("", false, "");
     return;
   }
